Fix DiscreteDistr.pdf to work with integer inputs.
When given ints as input, DiscreteDistr.pdf forces its output array to be an array of ints which coerces probabilities to integers and truncates them, returning unexpected output. This commit ensures the output array is an array of floats.

+++ pacal/distr.py
@@ -1050,7 +1050,7 @@ class DiscreteDistr(Distr):
         self.cumP = cumsum(self.pi)
     def pdf(self, x):
         """it override pdf() method to obtain discrete probabilities"""
-        yy = zeros_like(x)
+        yy = zeros_like(x, dtype=float)
         for i in range(len(self.pi)):
             yy[x==self.xi[i]] += float(self.pi[i])
         return yy
